+++ tests/test_z.py
@@ -3,8 +3,9 @@
 import numpy as np
 import tensorflow as tf
 
-# noinspection PyUnresolvedReferences
 from zfit import z
+# noinspection PyUnresolvedReferences
+from zfit.core.testing import setup_function, teardown_function, tester
 
 
 def numpy_func(x, a):