+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 Changelog
 *********
 
+
 Develop
 ========
 
@@ -25,11 +26,23 @@ Experimental
 Requirement changes
 -------------------
 
-
-
-
 Thanks
 ------
+
+
+0.5.6 (26.1.2020)
+=================
+
+Update to fix iminuit version
+
+Bug fixes and small changes
+---------------------------
+- Fix issue when using a `ComposedParameter` as the `rate` argument of a `Poisson` PDF
+
+Requirement changes
+-------------------
+- require iminuit < 2 to avoid breaking changes
+
 
 0.5.5 (20.10.2020)
 ==================